requirements.txt
streamlit>=1.45,<2.0
pandas>=2.2,<3.0
numpy>=1.26,<3.0
scikit-learn>=1.5,<2.0
pillow>=10.4,<12.0

